requirements.txt
discord.py>=2.3.2
pymongo[srv]
python-dotenv
Flask
